Refactor: move process_message to messaging_service
